chore: mover DoctorRepository para o pacote correto
